Move sidebar-related javascript into seperate file
