+++ assets/project-overview.js
@@ -1,3 +1,23 @@
+/**
+ * Project Overview bundle entrypoint.
+ *
+ * Subsystems that run on the project-overview page, in roughly the order they
+ * appear in this file:
+ *   - Filter pill init + select2 wiring (users / projects / other filters /
+ *     columns) and the cross-pill mutex with the "All" sentinel.
+ *   - View tabs widget (hidden in the redesigned UI, driven by sidebar clicks)
+ *     including URL pushState, unsaved-changes tracking, and lazy-loading the
+ *     active panel's table on activation.
+ *   - Ticket-row inline edits (status, priority, due date, assignee, tags,
+ *     hours, milestone) that PATCH the API and animate save success/error.
+ *   - Lazy-load sentinel for paginated row insertion.
+ *
+ * Sidebar view navigation has been extracted to `./sidebar.js`; the two
+ * functions imported below are the only cross-module touch points. New
+ * subsystems should also be extracted to their own module under `assets/`
+ * and wired in from here rather than appended to this file.
+ */
+
 import 'select2';
 import 'select2/dist/css/select2.css';
 import flatpickr from 'flatpickr';
@@ -6,6 +26,10 @@ import 'flatpickr/dist/flatpickr.min.css';
 import TomSelect from 'tom-select';
 import 'tom-select/dist/css/tom-select.bootstrap5.css';
 import './project-overview.css';
+import {
+  initSidebarViewNavigation,
+  applySidebarActiveState,
+} from './sidebar.js';
 
 $(document).ready(function () {
   window.frontendDateFormat = $(document).find('#frontendDateFormat').val();
@@ -156,106 +180,11 @@ function initSaveChangesSubmit() {
 }
 
 /**
- * Sidebar view navigation. The sidebar items are injected server-side by
- * `register.php::addProjectOverviewMenuPoint()` and carry `data-view-key` plus
- * htmx attrs that load the matching filter panel. We layer extra behaviour on
- * top:
- *
- *   - Click on a view item drives the (CSS-hidden) jQuery UI tabs widget so
- *     the existing activate handler runs — that's where URL pushState, form-
- *     state caching, `#selectedViewId` updates, save-button visibility, and
- *     panel show/hide all live. Reusing it keeps one source of truth.
- *   - Active class is mirrored from `#selectedViewId` onto the sidebar <li>.
- *   - Drag-reorder is scoped to plugin view items only; other sidebar items
- *     (Dashboard, Project Hub, etc.) stay put.
- *   - The 3-dot context-menu trigger inside each item opens the shared
- *     `#view-context-menu`.
- *
- * Only active on the ProjectOverview page (detected via the presence of
- * `#filtersContainer`). On other pages the sidebar links navigate naturally
- * via their `href`.
- */
-function initSidebarViewNavigation() {
-  if (!document.getElementById('filtersContainer')) return;
-
-  applySidebarActiveState();
-
-  document.body.addEventListener('click', function (e) {
-    // Context-menu trigger (a child of the <a>) — intercept before the link.
-    const trigger = e.target.closest('[data-context-trigger]');
-    if (trigger) {
-      const link = trigger.closest('a.projectoverview-view-item');
-      if (link) {
-        e.preventDefault();
-        e.stopPropagation();
-        openSidebarContextMenu(link, trigger);
-      }
-      return;
-    }
-
-    const link = e.target.closest('a.projectoverview-view-item');
-    if (!link) return;
-
-    const viewKey = link.getAttribute('data-view-key');
-    if (!viewKey) return;
-
-    // Drive the (hidden) tabs widget so its activate handler does the
-    // bookkeeping. htmx fires its own hx-get on this same click independently.
-    if (window.jQuery && window.jQuery.fn.tabs) {
-      const $tab = window.jQuery(
-        '#projectOverviewTabs li[data-target="' +
-          viewKey.replace(/"/g, '\\"') +
-          '"]'
-      );
-      const tabIndex = $tab.index();
-      if (tabIndex >= 0) {
-        try {
-          window
-            .jQuery('#projectOverviewTabs')
-            .tabs('option', 'active', tabIndex);
-        } catch (err) {
-          console.warn('[ProjectOverview] tabs() activation failed', err);
-        }
-      }
-    }
-
-    applySidebarActiveState();
-  });
-
-  // Move all saved-view <li>s into a scrollable wrapper so the sidebar can't
-  // grow unbounded when the user has many views. The "+ New view" item is
-  // deliberately left outside so it stays pinned at the bottom and visible.
-  // Must run before initSidebarSortable() since the sortable binds to the
-  // <ul> the items actually live in after this move.
-  wrapSidebarViewListInScroller();
-
-  initSidebarSortable();
-
-  // After the views live inside the scroll container, make sure the active
-  // one is visible — if the page was loaded with ?view=<id> pointing at an
-  // entry that sits past the scroller's fold, the user would otherwise see
-  // an empty/unrelated portion of the list.
-  scrollActiveSidebarViewIntoView();
-}
-
-/**
- * Scroll the currently-active sidebar view into the visible area of the
- * scrollable view list. No-op when the active item is already fully visible.
- */
-function scrollActiveSidebarViewIntoView() {
-  const active = document.querySelector('li.projectoverview-view-active');
-  if (!active) return;
-  // `block: 'nearest'` minimizes scroll: only moves when the element is out
-  // of the viewport of its nearest scrollable ancestor, which is our inner
-  // <ul.projectoverview-views-inner-list>.
-  active.scrollIntoView({ block: 'nearest' });
-}
-
-/**
- * Horizontal-tab-strip counterpart of {@link scrollActiveSidebarViewIntoView}.
- * Used on init so a page loaded with `?view=<id>` pointing at a tab past the
- * fold of the scrollable strip isn't hidden behind the sticky "+ Ny visning"
- * pin — the activate handler already scrolls on user-driven tab switches.
+ * Counterpart for the (CSS-hidden) horizontal tab strip — separate from the
+ * sidebar scroll helper that lives in `./sidebar.js`. Used on init so a page
+ * loaded with `?view=<id>` pointing at a tab past the fold of the scrollable
+ * strip isn't hidden behind the sticky "+ New view" pin — the activate handler
+ * already scrolls on user-driven tab switches.
  */
 function scrollActiveTabIntoView() {
   const active = document.querySelector(
@@ -263,183 +192,6 @@ function scrollActiveTabIntoView() {
   );
   if (!active) return;
   active.scrollIntoView({ inline: 'nearest', block: 'nearest' });
-}
-
-/**
- * Wraps the saved-view <li>s in a max-height + overflow:auto container so a
- * long view list scrolls inside the sidebar instead of pushing the layout.
- * Idempotent — does nothing if the wrapper already exists.
- */
-function wrapSidebarViewListInScroller() {
-  if (document.querySelector('.projectoverview-views-scroller')) return;
-
-  const firstView = document.querySelector('a.projectoverview-view-item');
-  if (!firstView) return;
-  const firstLi = firstView.closest('li');
-  const outerUl = firstLi ? firstLi.parentElement : null;
-  if (!outerUl) return;
-
-  // Collect every <li> whose anchor is a view item AND not the "+ New view"
-  // entry. Order is preserved by walking the outer <ul>'s children.
-  const viewLis = Array.from(outerUl.children).filter(function (li) {
-    const a = li.querySelector('a.projectoverview-view-item');
-    return a && !a.classList.contains('projectoverview-new-view-item');
-  });
-  if (viewLis.length === 0) return;
-
-  const wrapperLi = document.createElement('li');
-  wrapperLi.className = 'projectoverview-views-scroller';
-
-  const innerUl = document.createElement('ul');
-  innerUl.className = 'projectoverview-views-inner-list';
-  for (const li of viewLis) {
-    innerUl.appendChild(li);
-  }
-  wrapperLi.appendChild(innerUl);
-
-  // Insert the wrapper at the position the first view occupied (which is now
-  // adjacent to the "+ New view" item).
-  outerUl.insertBefore(wrapperLi, outerUl.firstChild);
-  // Move it to just before the "+ New view" item — outerUl.firstChild was the
-  // first remaining child after removing the views, which may not be the
-  // intended position. Recalculate explicitly.
-  const newViewLi = outerUl.querySelector(
-    'a.projectoverview-new-view-item'
-  )?.parentElement;
-  if (newViewLi) {
-    outerUl.insertBefore(wrapperLi, newViewLi);
-  }
-}
-
-/**
- * Mark the sidebar <li> whose link's `data-view-key` matches `#selectedViewId`
- * with `.projectoverview-view-active`, clearing it from all other view items.
- */
-function applySidebarActiveState() {
-  const selected = document.getElementById('selectedViewId');
-  const activeKey = selected ? selected.value : null;
-  document
-    .querySelectorAll('a.projectoverview-view-item')
-    .forEach(function (link) {
-      const li = link.parentElement;
-      if (!li) return;
-      li.classList.toggle(
-        'projectoverview-view-active',
-        link.getAttribute('data-view-key') === activeKey
-      );
-    });
-}
-
-/**
- * Open the shared #view-context-menu anchored to the trigger element (the
- * 3-dot span). Anchoring to the trigger instead of the whole <li> keeps the
- * menu visually attached to what was clicked.
- *
- * @param {HTMLAnchorElement} link     The sidebar <a> for the view.
- * @param {HTMLElement}       [anchor] Element to position the menu next to.
- *                                     Falls back to the trigger inside `link`,
- *                                     then to the link itself.
- */
-function openSidebarContextMenu(link, anchor) {
-  const viewId = link.getAttribute('data-view-key');
-  if (!viewId || viewId === '__new') return;
-
-  const contextMenu = document.getElementById('view-context-menu');
-  if (!contextMenu) return;
-
-  const isSubscription =
-    link.getAttribute('data-is-subscription') === 'true' ||
-    link.getAttribute('data-is-transient-subscription') === 'true';
-  const subscribeToken = link.getAttribute('data-subscribe-token') || '';
-  const labelSpan = link.querySelector('.view-label');
-  const currentName = labelSpan ? labelSpan.textContent.trim() : '';
-  // #view-context-menu is position: fixed, so top/left are viewport-relative.
-  // Left anchors to the main content area so the popup sits outside the
-  // sidebar at a predictable column. Top mirrors the clicked view <li>'s
-  // viewport y so the popup appears at the same vertical as the clicked row.
-  const mainContent =
-    document.querySelector('.maincontentinner') ||
-    document.querySelector('.maincontent');
-  const mainContentLeft = mainContent
-    ? mainContent.getBoundingClientRect().left
-    : (anchor || link).getBoundingClientRect().right + 4;
-  const rowTop = link.parentElement
-    ? link.parentElement.getBoundingClientRect().top
-    : (anchor || link).getBoundingClientRect().top;
-
-  contextMenu.dataset.mode = isSubscription ? 'subscription' : 'owned';
-  contextMenu.style.left = mainContentLeft + 'px';
-  contextMenu.style.top = rowTop + 'px';
-  contextMenu.classList.add('shown');
-
-  const title = contextMenu.querySelector('#contextMenuTitle');
-  if (title) title.textContent = currentName;
-  const setVal = function (sel, v) {
-    const el = contextMenu.querySelector(sel);
-    if (el) el.value = v;
-  };
-  setVal('input[name="viewName"]', currentName);
-  setVal('input[name="view"]', viewId);
-  setVal('input[name="subscribeToken"]', subscribeToken);
-
-  if (!isSubscription) {
-    requestAnimationFrame(function () {
-      const input = contextMenu.querySelector('input[name="viewName"]');
-      if (input) input.focus();
-    });
-  }
-}
-
-/**
- * Apply jQuery UI Sortable to the sidebar <ul> containing the view items.
- * The `items` selector restricts dragging to plugin view items only — other
- * menu entries (Dashboard, Project Hub, etc.) and the "+ New view" item stay
- * fixed in place.
- */
-function initSidebarSortable() {
-  const firstViewItem = document.querySelector('a.projectoverview-view-item');
-  if (!firstViewItem) return;
-  const ul = firstViewItem.closest('ul');
-  if (!ul) return;
-  if (!window.jQuery || !window.jQuery.fn.sortable) return;
-
-  window.jQuery(ul).sortable({
-    items:
-      'li:has(a.projectoverview-view-item):not(:has(a.projectoverview-new-view-item))',
-    axis: 'y',
-    tolerance: 'pointer',
-    handle: 'a.projectoverview-view-item',
-    // Without forcePlaceholderSize, the placeholder collapses to its content
-    // height (effectively nothing for an empty <li>), making the drop target
-    // ambiguous. Pairing it with a custom placeholder class lets us match the
-    // height of the dragged item.
-    placeholder: 'projectoverview-view-placeholder',
-    forcePlaceholderSize: true,
-    // Disambiguate click vs. drag: the cursor must move ~5px AND the press has
-    // to last ~150ms before the drag engages. Without these, jQuery UI starts
-    // a drag on any pixel of mouse movement, so quick clicks with slight
-    // wobble get hijacked into drags and the click action never fires.
-    delay: 150,
-    distance: 5,
-    update: function () {
-      const newOrder = Array.from(
-        ul.querySelectorAll(
-          'a.projectoverview-view-item:not(.projectoverview-new-view-item)'
-        )
-      ).map(function (a) {
-        return a.getAttribute('data-view-key');
-      });
-      window.jQuery.ajax({
-        dataType: 'json',
-        url: '/ProjectOverview/ProjectOverview/post',
-        method: 'POST',
-        data: { action: 'saveTabOrder', order: newOrder },
-        error: function (xhr, status, error) {
-          console.error('[ProjectOverview] saveTabOrder failed', error);
-        },
-      });
-    },
-  });
 }
 
 /**
